Rework Code
Removed FavoriteActivity and connected activities.
DetailActivity now will display favorites that are stored in a local database.
Added LoadDataService that manages insertion and deletions from the database.
DetailActivity implements a LocalBroadcastReceiver to manage response from the LoadDataService.
Made changes to the datamodel the MovieClass now has an addtional field - byte [] poster where a copy of the poster is stored for the favorites listing.